--- a/cargo-near-build/src/near/abi/mod.rs
+++ b/cargo-near-build/src/near/abi/mod.rs
@@ -91,11 +91,6 @@ pub fn write_to_file(
                 "skipped wrting file `{}` on identical contents",
                 out_path_abi,
             );
-
-            println!(
-                "skipped wrting file `{}` on identical contents",
-                out_path_abi,
-            );
             return Ok(abi_types::Result { path: out_path_abi });
         }
     }