--- a/cargo-near-build/src/near/build/mod.rs
+++ b/cargo-near-build/src/near/build/mod.rs
@@ -57,7 +57,7 @@ pub fn run(args: Opts) -> eyre::Result<CompilationArtifact> {
 
     let out_dir = crate_metadata.resolve_output_dir(args.out_dir.clone().map(Into::into))?;
 
-    let mut cargo_args = vec!["--target", COMPILATION_TARGET];
+    let mut cargo_args = vec!["--verbose", "--target", COMPILATION_TARGET];
     let cargo_feature_args = {
         let mut feat_args = vec![];
         if let Some(features) = args.features.as_ref() {