--- a/cargo-near-build/src/near/build/mod.rs
+++ b/cargo-near-build/src/near/build/mod.rs
@@ -216,6 +216,7 @@ fn maybe_wasm_opt_step(
         pretty_print::handle_step(
             "Running an optimize for size post-step with wasm-opt...",
             || {
+                let start = std::time::Instant::now();
                 println!(
                     "{} -> {}",
                     format!("{}", input_path).cyan(),
@@ -223,6 +224,7 @@ fn maybe_wasm_opt_step(
                 );
                 wasm_opt::OptimizationOptions::new_optimize_for_size()
                     .run(input_path, opt_destination.path())?;
+                pretty_print::duration_millis(start, "wasm-opt -O");
                 Ok(())
             },
         )?;