@dj8yfo 2025-01-17
adding https://github.com/dj8yfo/neardevhub-contract/commit/cbbf0ee589f45e2174e2b264ea3c7a869cfdbc7f
prevents partial rebuilds of `neardevhub-contract` and `neardevhub-community-factory` on 

```bash
 │        Dirty near-sdk-macros v5.7.0: the environment variable CARGO_NEAR_ABI_PATH changed
 │    Compiling near-sdk-macros v5.7.0
```
due to `near-sdk-macros` crate's builds being shared in the same release profile. 

Only wasm-opt computing overhead is left per each test with `cargo nextest run` (5 s for neardevhub and 1 s for community-factory)