@dj8yfo 2025-01-20
the result of adding https://github.com/near/cargo-near/pull/292/commits/e7466996403207d5d59b02c6f7b35c0de0f4b22b is measurable: 

update rev (normal verbosity, wasm-opt duration, wasm-opt not skipped on rerun)
https://github.com/dj8yfo/neardevhub-contract/commit/c34fff39196c3704a0e3bf339d541086252d31e2
  - cargo test 
    - [ ] 56.72 s
    - [ ] 54.13 s
    - [ ] 54.34 s
  - cargo nextest run 
    - [ ] 42.6 s
    - [ ] 42.6 s
    - [ ] 43.7 s

update rev of cargo-near-build patch (wasm-opt stage skipped)
https://github.com/dj8yfo/neardevhub-contract/commit/3e03d0e4d2d3264ed586d3df5f62298b8a4f4613
  - cargo test 
    - [ ] 33 s
    - [ ] 32.43 s
    - [ ] 31.87 s
  - cargo nextest run 
    - [ ] 32.25 s
    - [ ] 32.35 s
    - [ ] 32.31 s