@dj8yfo 2025-01-20
running `cargo test` on clean state (taking into account inital contracts' build) takes 542 s,
and `cargo nextest run` - 472 s 

`cargo nextest run` from scratch on the same code with unpatched version (from crates.io and not
from this pr) https://github.com/dj8yfo/neardevhub-contract/commit/dd3ada135ef57e6bd336894840db6550d01b738f
of cargo-near-build - 527 s, so the optimization of this pr is 55 s for 7 tests which build contracts.
And 72 s unpatched vs 32 s patched for second run for 7 tests.